pyproject.toml
[project]
name = "RaySR"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.*"  # Tensorflow 2.18.1 does not provide wheels for Python 3.13

authors = [
    { name = "Simon Wanna", email = "simonwanna@gmail.com" }
]

dependencies = [
    "ruff>=0.12.2",
    "pre-commit>=4.3.0",
    "sionna>=1.1.0",
    "matplotlib>=3.10.6",
    "torch==2.7.1",
    "lightning>=2.5.5",
    "tqdm>=4.67.1",
    "hydra-core>=1.3.2",
    "wandb>=0.22.0",
]

[dependency-groups]
dev = ["ipykernel>=6.30.1"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "ANN001",   # missing type annotation for function argument
    "ANN201",   # missing return type annotation for public function
    "I001",     # isort
]

[tool.ruff]
exclude = ["src/experiments/*", "*.ipynb"]
line-length = 120

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/poc"]

[project.scripts]
generate = "poc.generate_data:generate_data"
train = "poc.train:train"
test = "poc.test:test"